+++ tulip/shared/polyfills.h
@@ -19,7 +19,7 @@ extern void esp32s3_display_stop();
 uint8_t rand_uint8();
 void take_semaphore();
 int64_t get_time_ms();
-int64_t get_ticks_ms();
+int32_t get_ticks_ms();
 int64_t get_time_us();
 void * malloc_caps(uint32_t size, uint32_t flags);
 void *calloc_caps(uint32_t align, uint32_t count, uint32_t size, uint32_t flags);
